Refactor FHIR compactor
