+++ src/fhir_mcp_server/compactor/types/annotation.py
