+++ src/fhir_mcp_server/compactor/types/attachment.py
