+++ src/fhir_mcp_server/compactor/types/base.py
