--- src/fhir_mcp_server/compactor/types/complex_types/__init__.py
@@ -1,52 +1,0 @@
-# Copyright (c) 2026, WSO2 LLC. (https://www.wso2.com/) All Rights Reserved.
-
-# WSO2 LLC. licenses this file to you under the Apache License,
-# Version 2.0 (the "License"); you may not use this file except
-# in compliance with the License.
-# You may obtain a copy of the License at
-
-# http://www.apache.org/licenses/LICENSE-2.0
-
-# Unless required by applicable law or agreed to in writing,
-# software distributed under the License is distributed on an
-# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-# KIND, either express or implied. See the License for the
-# specific language governing permissions and limitations
-# under the License.
-
-
-from .address import Address
-from .annotation import Annotation
-from .attachment import Attachment
-from .codeable_concept import CodeableConcept
-from .coding import Coding
-from .contact_point import ContactPoint
-from .extension import Extension
-from .human_name import HumanName
-from .identifier import Identifier
-from .money import Money
-from .period import Period
-from .quantity import Quantity
-from .range import Range
-from .ratio import Ratio
-from .reference import Reference
-from .timing import Timing
-
-__all__ = [
-    "Address",
-    "Annotation",
-    "Attachment",
-    "CodeableConcept",
-    "Coding",
-    "ContactPoint",
-    "Extension",
-    "HumanName",
-    "Identifier",
-    "Money",
-    "Period",
-    "Quantity",
-    "Range",
-    "Ratio",
-    "Reference",
-    "Timing",
-]
