+++ tests/unit/test_fhir_compactor.py
@@ -15,11 +15,33 @@
 # under the License.
 
 import pytest
+from datetime import date, datetime
 from fhir_mcp_server.compactor.dispatch import compact_resource
+from fhir_mcp_server.compactor.types import (
+    Address,
+    Annotation,
+    Attachment,
+    CodeableConcept,
+    Coding,
+    ContactPoint,
+    Extension,
+    HumanName,
+    Identifier,
+    Money,
+    Period,
+    Quantity,
+    Range,
+    Ratio,
+    Reference,
+    Timing,
+)
 
 
 class TestCompactCodeableConcept:
+    """Test compact_resource with CodeableConcept payloads."""
+
     def test_uses_text_when_present(self):
+        """Test that text field takes precedence over coding."""
         v = {
             "coding": [
                 {
@@ -33,12 +55,14 @@ class TestCompactCodeableConcept:
         assert compact_resource(v) == "Body Height"
 
     def test_text_only_no_coding(self):
+        """Test that text-only CodeableConcept compacts to the text value."""
         assert (
             compact_resource({"text": "uncoded free text result"})
             == "uncoded free text result"
         )
 
     def test_uses_display_and_code_when_no_text(self):
+        """Test that display and code are combined when text is absent."""
         v = {
             "coding": [
                 {
@@ -51,22 +75,27 @@ class TestCompactCodeableConcept:
         assert compact_resource(v) == "Body Height (8302-2)"
 
     def test_display_only_no_code(self):
+        """Test that display alone is used when code is absent."""
         v = {"coding": [{"display": "Body Height"}]}
         assert compact_resource(v) == "Body Height"
 
     def test_uses_system_and_code_when_no_display(self):
+        """Test that system and code are combined when display is absent."""
         v = {"coding": [{"system": "http://loinc.org", "code": "8302-2"}]}
         assert compact_resource(v) == "http://loinc.org|8302-2"
 
     def test_code_only_no_system_no_display(self):
+        """Test that bare code is returned when system and display are absent."""
         v = {"coding": [{"code": "8302-2"}]}
         assert compact_resource(v) == "8302-2"
 
     def test_empty_coding_returns_raw(self):
+        """Test that unresolvable coding returns the raw dict."""
         v = {"coding": [{"userSelected": True}]}
         assert compact_resource(v) == v
 
     def test_multiple_codings_uses_first(self):
+        """Test that only the first coding entry is used."""
         v = {
             "coding": [
                 {
@@ -81,60 +110,80 @@ class TestCompactCodeableConcept:
 
 
 class TestCompactCoding:
+    """Test compact_resource with Coding payloads."""
+
     def test_uses_display_and_code(self):
+        """Test that display and code are combined."""
         v = {"system": "http://loinc.org", "code": "8302-2", "display": "Body Height"}
         assert compact_resource(v) == "Body Height (8302-2)"
 
     def test_display_only(self):
+        """Test that display alone is returned when code is absent."""
         v = {"display": "Body Height"}
         assert compact_resource(v) == "Body Height"
 
     def test_uses_system_and_code_when_no_display(self):
+        """Test that system and code are combined when display is absent."""
         v = {"system": "http://loinc.org", "code": "8302-2"}
         assert compact_resource(v) == "http://loinc.org|8302-2"
 
     def test_code_only(self):
+        """Test that bare code is returned when system and display are absent."""
         v = {"code": "8302-2"}
         assert compact_resource(v) == "8302-2"
 
 
 class TestCompactQuantity:
+    """Test compact_resource with Quantity payloads."""
+
     def test_value_and_unit(self):
+        """Test that value and unit are joined with a space."""
         assert compact_resource({"value": 7.2, "unit": "mmol/L"}) == "7.2 mmol/L"
 
     def test_with_comparator(self):
+        """Test that comparator is prepended to the value."""
         assert (
             compact_resource({"value": 5.0, "comparator": ">=", "unit": "mg/dL"})
             == ">=5 mg/dL"
         )
 
     def test_integer_value_no_trailing_zeros(self):
+        """Test that integer values are rendered without decimal places."""
         assert compact_resource({"value": 150, "unit": "cm"}) == "150 cm"
 
     def test_value_only_no_unit(self):
+        """Test that value alone is returned when unit is absent."""
         assert (
             compact_resource({"value": 42, "system": "http://unitsofmeasure.org"})
             == "42"
         )
 
     def test_unit_only_no_value(self):
+        """Test that unit alone is returned when value is absent."""
         assert compact_resource({"unit": "cm"}) == "cm"
 
 
 class TestCompactQuantitySubtypes:
+    """Test compact_resource with Quantity subtype payloads (Age, Count, Distance, etc.)."""
+
     def test_age(self):
+        """Test Age subtype compaction."""
         assert compact_resource({"value": 30, "unit": "yr"}) == "30 yr"
 
     def test_count(self):
+        """Test Count subtype compaction."""
         assert compact_resource({"value": 3, "unit": "{count}"}) == "3 {count}"
 
     def test_distance(self):
+        """Test Distance subtype compaction."""
         assert compact_resource({"value": 5, "unit": "km"}) == "5 km"
 
     def test_duration(self):
+        """Test Duration subtype compaction."""
         assert compact_resource({"value": 30, "unit": "min"}) == "30 min"
 
     def test_money_quantity(self):
+        """Test MoneyQuantity subtype compaction."""
         assert (
             compact_resource(
                 {"value": 100, "unit": "USD", "system": "urn:iso:std:iso:4217"}
@@ -143,11 +192,15 @@ class TestCompactQuantitySubtypes:
         )
 
     def test_simple_quantity(self):
+        """Test SimpleQuantity subtype compaction."""
         assert compact_resource({"value": 5, "unit": "mg"}) == "5 mg"
 
 
 class TestCompactRange:
+    """Test compact_resource with Range payloads."""
+
     def test_low_and_high(self):
+        """Test that low and high bounds are joined with an en-dash."""
         v = {
             "low": {"value": 3.5, "unit": "mmol/L"},
             "high": {"value": 5.5, "unit": "mmol/L"},
@@ -155,38 +208,36 @@ class TestCompactRange:
         assert compact_resource(v) == "3.5 mmol/L – 5.5 mmol/L"
 
     def test_only_low(self):
+        """Test that only the low bound is returned when high is absent."""
         v = {"low": {"value": 3.5, "unit": "mmol/L"}}
         assert compact_resource(v) == "3.5 mmol/L"
 
     def test_only_high(self):
+        """Test that only the high bound is returned when low is absent."""
         v = {"high": {"value": 5.5, "unit": "mmol/L"}}
         assert compact_resource(v) == "5.5 mmol/L"
 
     def test_equal_bounds_allowed(self):
-        from fhir_mcp_server.compactor.types.complex_types.range import Range
-
+        """Test that equal low and high bounds are accepted."""
         r = Range.model_validate({"low": {"value": 5.0}, "high": {"value": 5.0}})
         assert r.low.value == r.high.value
 
 
 class TestPeriodValidation:
-    def test_date_objects_accepted(self):
-        from datetime import date
-        from fhir_mcp_server.compactor.types.complex_types.period import Period
+    """Test Period model validation with various date/datetime inputs."""
 
+    def test_date_objects_accepted(self):
+        """Test that Python date objects are accepted as period boundaries."""
         p = Period.model_validate({"start": date(2024, 1, 1), "end": date(2024, 6, 1)})
         assert p.start == date(2024, 1, 1)
 
     def test_equal_start_end_allowed(self):
-        from fhir_mcp_server.compactor.types.complex_types.period import Period
-
+        """Test that equal start and end dates are accepted."""
         p = Period.model_validate({"start": "2024-01-01", "end": "2024-01-01"})
         assert p.start == p.end
 
     def test_period_datetime_validation(self):
-        from fhir_mcp_server.compactor.types.complex_types.period import Period
-        from datetime import datetime
-
+        """Test that Python datetime objects are accepted as period boundaries."""
         p = Period.model_validate(
             {"start": datetime(2024, 1, 1, 10), "end": datetime(2024, 1, 1, 12)}
         )
@@ -194,7 +245,10 @@ class TestPeriodValidation:
 
 
 class TestCompactRatio:
+    """Test compact_resource with Ratio payloads."""
+
     def test_numerator_and_denominator(self):
+        """Test that numerator and denominator are joined with a slash."""
         v = {
             "numerator": {"value": 1, "unit": "mg"},
             "denominator": {"value": 10, "unit": "mL"},
@@ -202,58 +256,75 @@ class TestCompactRatio:
         assert compact_resource(v) == "1 mg/10 mL"
 
     def test_missing_denominator_returns_raw(self):
+        """Test that a missing denominator returns the raw dict."""
         v = {"numerator": {"value": 1, "unit": "mg"}}
         assert compact_resource(v) == v
 
     def test_missing_numerator_returns_raw(self):
+        """Test that a missing numerator returns the raw dict."""
         v = {"denominator": {"value": 10, "unit": "mL"}}
         assert compact_resource(v) == v
 
 
 class TestCompactPeriod:
+    """Test compact_resource with Period payloads."""
+
     def test_start_and_end(self):
+        """Test that start and end are joined with an en-dash."""
         assert (
             compact_resource({"start": "2024-01-01", "end": "2024-06-30"})
             == "2024-01-01 – 2024-06-30"
         )
 
     def test_only_start(self):
+        """Test that a start-only period is prefixed with 'from'."""
         assert compact_resource({"start": "2024-01-01"}) == "from 2024-01-01"
 
     def test_only_end(self):
+        """Test that an end-only period is prefixed with 'until'."""
         assert compact_resource({"end": "2024-06-30"}) == "until 2024-06-30"
 
 
 class TestCompactHumanName:
+    """Test compact_resource with HumanName payloads."""
+
     def test_full_name_with_prefix(self):
+        """Test that prefix, given, and family are joined in order."""
         v = {"use": "official", "family": "Smith", "given": ["John"], "prefix": ["Dr."]}
         assert compact_resource(v) == "Dr. John Smith"
 
     def test_nickname_appends_use(self):
+        """Test that non-official use values are appended in parentheses."""
         v = {"use": "nickname", "given": ["Johnny"]}
         assert compact_resource(v) == "Johnny (nickname)"
 
     def test_uses_text_when_present(self):
+        """Test that the text field takes precedence over structured name parts."""
         v = {"text": "John Smith", "family": "Smith", "given": ["John"]}
         assert compact_resource(v) == "John Smith"
 
     def test_suffix_appended(self):
+        """Test that suffix is appended after the family name."""
         v = {"family": "Smith", "given": ["John"], "suffix": ["Jr."]}
         assert compact_resource(v) == "John Smith Jr."
 
     def test_official_use_suppressed(self):
+        """Test that 'official' use is not appended to the name."""
         v = {"use": "official", "family": "Smith", "given": ["John"]}
         assert compact_resource(v) == "John Smith"
 
     def test_maiden_use_shown(self):
+        """Test that 'maiden' use is appended in parentheses."""
         v = {"use": "maiden", "family": "Jones", "given": ["Jane"]}
         assert compact_resource(v) == "Jane Jones (maiden)"
 
     def test_usual_use_shown(self):
+        """Test that 'usual' use is appended in parentheses."""
         v = {"use": "usual", "family": "Smith", "given": ["Johnny"]}
         assert compact_resource(v) == "Johnny Smith (usual)"
 
     def test_multiple_names_compacted_as_list(self):
+        """Test that a list of HumanName entries is compacted to a list of strings."""
         v = [
             {
                 "use": "official",
@@ -272,7 +343,10 @@ class TestCompactHumanName:
 
 
 class TestCompactAddress:
+    """Test compact_resource with Address payloads."""
+
     def test_full_address(self):
+        """Test that all address fields are formatted correctly."""
         v = {
             "line": ["123 Main St"],
             "city": "Boston",
@@ -283,11 +357,13 @@ class TestCompactAddress:
         assert compact_resource(v) == "123 Main St, Boston MA 02101, US"
 
     def test_uses_text_when_present(self):
+        """Test that the text field takes precedence over structured address parts."""
         # text + city ensures HumanName validation fails (no city field) and Address is reached
         v = {"text": "123 Main St, Boston MA 02101", "city": "Boston"}
         assert compact_resource(v) == "123 Main St, Boston MA 02101"
 
     def test_district_included_in_location(self):
+        """Test that district is included between city and state."""
         v = {
             "city": "Boston",
             "district": "Suffolk",
@@ -297,44 +373,55 @@ class TestCompactAddress:
         assert compact_resource(v) == "Boston Suffolk MA 02101"
 
     def test_use_and_type_appended(self):
+        """Test that use and type are appended in parentheses."""
         v = {"use": "home", "type": "postal", "line": ["123 Main St"], "city": "Boston"}
         assert compact_resource(v) == "123 Main St, Boston (home, postal)"
 
     def test_use_only(self):
+        """Test that use alone is appended in parentheses when type is absent."""
         v = {"use": "work", "city": "Boston"}
         assert compact_resource(v) == "Boston (work)"
 
 
 class TestCompactContactPoint:
+    """Test compact_resource with ContactPoint payloads."""
+
     def test_phone_with_use(self):
+        """Test that system, value, and use are formatted correctly."""
         assert (
             compact_resource({"system": "phone", "value": "555-1234", "use": "home"})
             == "phone: 555-1234 (home)"
         )
 
     def test_email_without_use(self):
+        """Test that system and value are formatted without use when absent."""
         assert (
             compact_resource({"system": "email", "value": "john@example.com"})
             == "email: john@example.com"
         )
 
     def test_rank_appended(self):
+        """Test that rank is appended with a # prefix."""
         v = {"system": "phone", "value": "555-1234", "use": "home", "rank": 1}
         assert compact_resource(v) == "phone: 555-1234 (home) #1"
 
 
 class TestCompactIdentifier:
+    """Test compact_resource with Identifier payloads."""
+
     def test_system_value_and_use(self):
+        """Test that system, value, and use are formatted correctly."""
         v = {"system": "http://hospital.org/mrn", "value": "MRN123", "use": "official"}
         assert compact_resource(v) == "http://hospital.org/mrn|MRN123 [official]"
 
     def test_not_confused_with_contact_point(self):
-        # URI system should not be detected as ContactPoint
+        """Test that a URI system is not mistaken for a ContactPoint."""
         v = {"system": "http://hospital.org/mrn", "value": "MRN123"}
         result = compact_resource(v)
         assert "http://hospital.org/mrn" in result
 
     def test_type_prepended_as_label(self):
+        """Test that the type text is prepended as a label."""
         v = {
             "type": {"text": "MRN"},
             "system": "http://hospital.org/mrn",
@@ -343,9 +430,11 @@ class TestCompactIdentifier:
         assert compact_resource(v) == "MRN: http://hospital.org/mrn|MRN123"
 
     def test_value_only(self):
+        """Test that value alone is returned when system is absent."""
         assert compact_resource({"value": "MRN123"}) == "MRN123"
 
     def test_system_only_returns_entire_json(self):
+        """Test that a system-only identifier returns the raw dict."""
         assert (
             compact_resource({"system": "http://hospital.org/mrn"})
             == {"system": "http://hospital.org/mrn"}
@@ -353,23 +442,30 @@ class TestCompactIdentifier:
 
 
 class TestCompactAttachment:
+    """Test compact_resource with Attachment payloads."""
+
     def test_title_and_content_type(self):
+        """Test that title and contentType are combined."""
         v = {"title": "Discharge Summary", "contentType": "application/pdf"}
         assert compact_resource(v) == "Discharge Summary (application/pdf)"
 
     def test_title_without_content_type(self):
+        """Test that title alone is returned when contentType is absent."""
         v = {"title": "Discharge Summary"}
         assert compact_resource(v) == "Discharge Summary"
 
     def test_falls_back_to_url(self):
+        """Test that URL is used when title is absent."""
         v = {"url": "http://example.com/image.png", "contentType": "image/png"}
         assert compact_resource(v) == "http://example.com/image.png"
 
     def test_content_type_only(self):
+        """Test that contentType alone is returned when title and URL are absent."""
         v = {"contentType": "application/pdf"}
         assert compact_resource(v) == "application/pdf"
 
     def test_inline_text_data_decoded(self):
+        """Test that base64-encoded text data is decoded and returned as a string."""
         import base64
 
         v = {
@@ -379,6 +475,7 @@ class TestCompactAttachment:
         assert compact_resource(v) == "hello world"
 
     def test_inline_binary_data_as_data_uri(self):
+        """Test that base64-encoded binary data is returned as a data URI."""
         import base64
 
         raw = b"\x89PNG\r\n"
@@ -389,8 +486,7 @@ class TestCompactAttachment:
         )
 
     def test_hash_bytes_passthrough(self):
-        from fhir_mcp_server.compactor.types.complex_types.attachment import Attachment
-
+        """Test that raw bytes are accepted and stored as-is for hash."""
         raw_hash = b"\xde\xad\xbe\xef"
         a = Attachment.model_validate(
             {"contentType": "application/pdf", "hash": raw_hash}
@@ -398,15 +494,13 @@ class TestCompactAttachment:
         assert a.hash == raw_hash
 
     def test_data_bytes_passthrough(self):
-        from fhir_mcp_server.compactor.types.complex_types.attachment import Attachment
-
+        """Test that raw bytes are accepted and stored as-is for data."""
         raw = b"raw bytes"
         a = Attachment.model_validate({"contentType": "application/pdf", "data": raw})
         assert a.data == raw
 
     def test_attachment_hash_data_non_string(self):
-        from fhir_mcp_server.compactor.types.complex_types.attachment import Attachment
-
+        """Test that dict values for hash and data are rejected."""
         with pytest.raises(Exception):
             Attachment.model_validate(
                 {"contentType": "text/plain", "hash": {"dict": 1}, "data": {"dict": 2}}
@@ -414,19 +508,25 @@ class TestCompactAttachment:
 
 
 class TestCompactAnnotation:
+    """Test compact_resource with Annotation payloads."""
+
     def test_text_only(self):
+        """Test that text alone is returned as-is."""
         v = {"text": "Patient was fasting"}
         assert compact_resource(v) == "Patient was fasting"
 
     def test_text_with_time(self):
+        """Test that time is appended in parentheses."""
         v = {"text": "Patient was fasting", "time": "2024-01-01T09:00:00Z"}
         assert compact_resource(v) == "Patient was fasting (2024-01-01T09:00:00Z)"
 
     def test_text_with_author_string(self):
+        """Test that authorString is appended in parentheses."""
         v = {"text": "Patient was fasting", "authorString": "Dr. Smith"}
         assert compact_resource(v) == "Patient was fasting (Dr. Smith)"
 
     def test_text_with_author_and_time(self):
+        """Test that author and time are both appended, comma-separated."""
         v = {
             "text": "Patient was fasting",
             "authorString": "Dr. Smith",
@@ -435,6 +535,7 @@ class TestCompactAnnotation:
         assert compact_resource(v) == "Patient was fasting (Dr. Smith, 2024-01-01)"
 
     def test_author_reference(self):
+        """Test that authorReference is resolved to its reference string."""
         v = {
             "text": "I don't think this is true",
             "authorReference": {"reference": "Patient/example"},
@@ -447,15 +548,22 @@ class TestCompactAnnotation:
 
 
 class TestCompactMoney:
+    """Test compact_resource with Money payloads."""
+
     def test_value_and_currency(self):
+        """Test that value and currency are joined with a space."""
         assert compact_resource({"value": 49.99, "currency": "USD"}) == "49.99 USD"
 
     def test_value_without_currency(self):
+        """Test that value alone is returned when currency is absent."""
         assert compact_resource({"value": 49.99, "currency": None}) == "49.99"
 
 
 class TestCompactResourceList:
+    """Test compact_resource list handling and recursive dispatch behavior."""
+
     def test_compacts_each_item_in_list(self):
+        """Test that each item in a list is individually compacted."""
         data = [
             {"coding": [{"code": "M", "display": "Married"}]},
             {"coding": [{"code": "S", "display": "Single"}]},
@@ -463,11 +571,13 @@ class TestCompactResourceList:
         assert compact_resource(data) == ["Married (M)", "Single (S)"]
 
     def test_leaves_primitives_unchanged(self):
+        """Test that primitive values are returned unchanged."""
         assert compact_resource("hello") == "hello"
         assert compact_resource(42) == 42
         assert compact_resource(True) is True
 
     def test_recurses_into_unrecognized_dict(self):
+        """Test that unrecognized dicts are recursed into and their values compacted."""
         data = {
             "resourceType": "Patient",
             "valueQuantity": {"value": 7.2, "unit": "cm"},
@@ -479,6 +589,8 @@ class TestCompactResourceList:
 
 
 class TestCompactExtension:
+    """Test compact_resource with Extension payloads (flat and nested)."""
+
     # us-core-interpreter-needed — flat valueCoding extension
     INTERPRETER_NEEDED = {
         "url": "http://hl7.org/fhir/us/core/StructureDefinition/us-core-interpreter-needed",
@@ -559,6 +671,7 @@ class TestCompactExtension:
     }
 
     def test_flat_extension_returns_url_pipe_value(self):
+        """Test that a flat extension compacts to url|value."""
         result = compact_resource(self.INTERPRETER_NEEDED)
         assert (
             result
@@ -566,6 +679,7 @@ class TestCompactExtension:
         )
 
     def test_nested_extension_returns_dict_keyed_by_sub_url(self):
+        """Test that nested extension compacts to a dict keyed by sub-extension URLs."""
         result = compact_resource(self.RACE)
         assert isinstance(result, dict)
         assert (
@@ -575,6 +689,7 @@ class TestCompactExtension:
         assert result["text"] == "Mixed"
 
     def test_repeated_sub_url_becomes_list(self):
+        """Test that repeated sub-extension URLs are collected into a list."""
         result = compact_resource(self.RACE)
         assert isinstance(result["ombCategory"], list)
         assert result["ombCategory"] == [
@@ -586,6 +701,7 @@ class TestCompactExtension:
         assert result["detailed"] == ["Shoshone (1586-7)", "Filipino (2036-2)"]
 
     def test_nested_extension_with_codeable_concept_and_boolean(self):
+        """Test that valueCodeableConcept and valueBoolean are compacted in nested extensions."""
         result = compact_resource(self.TRIBAL_AFFILIATION)
         assert isinstance(result, dict)
         assert (
@@ -596,11 +712,13 @@ class TestCompactExtension:
         assert result["isEnrolled"] == "False"
 
     def test_unresolvable_nested_extension_returns_raw(self):
+        """Test that a nested extension with no sub-extensions returns the raw dict."""
         data = {"url": "http://example.org/ext", "extension": []}
         result = compact_resource(data)
         assert result == data
 
     def test_sub_extension_value_that_compacts_to_dict_is_dropped(self):
+        """Test that sub-extensions whose value cannot be reduced to a string are skipped."""
         # valueX that compact_resource cannot reduce to a string (unrecognized dict)
         # _extract_extension_value returns "" → sub-extension is skipped
         data = {
@@ -616,50 +734,61 @@ class TestCompactExtension:
         assert "unknown" not in result
 
     def test_float_value_in_extension(self):
+        """Test that float values are compacted to url|value."""
         data = {"url": "http://example.org/ext", "valueDecimal": 3.14}
         assert compact_resource(data) == "http://example.org/ext|3.14"
 
     def test_int_value_in_extension(self):
+        """Test that integer values are compacted to url|value."""
         data = {"url": "http://example.org/ext", "valueInteger": 42}
         assert compact_resource(data) == "http://example.org/ext|42"
 
     def test_no_value_keys_returns_raw(self):
+        """Test that an extension with no value key returns the raw dict."""
         data = {"url": "http://example.org/ext"}
         assert compact_resource(data) == data
 
     def test_multiple_value_keys_returns_raw(self):
+        """Test that an extension with multiple value keys returns the raw dict."""
         data = {"url": "http://example.org/ext", "valueString": "a", "valueInteger": 1}
         assert compact_resource(data) == data
 
     def test_extract_extension_value_list(self):
-        from fhir_mcp_server.compactor.types.complex_types.extension import Extension
-
+        """Test that a list-typed value field returns an empty string."""
         ext = Extension.model_validate({"url": "x", "valueList": [1, 2]})
         assert ext._extract_extension_value() == ""
 
     def test_compact_extension_nested_not_dict(self):
+        """Test that non-dict items in extension list return the raw value."""
         v = {"url": "x", "extension": ["string"]}
         assert compact_resource(v) == v
 
 
 class TestCompactTiming:
+    """Test compact_resource with Timing payloads."""
+
     def test_code_text_used_verbatim(self):
+        """Test that code text is returned as-is."""
         v = {"code": {"text": "Take medication in the morning on weekends"}}
         assert compact_resource(v) == "Take medication in the morning on weekends"
 
     def test_known_abbreviation_code(self):
+        """Test that known timing abbreviation codes are returned as-is."""
         v = {"code": {"coding": [{"code": "BID"}]}}
         assert compact_resource(v) == "BID"
 
     def test_every_8_hours(self):
+        """Test that frequency=1 over 8h period compacts to 'every 8h'."""
         v = {"repeat": {"frequency": 1, "period": 8, "periodUnit": "h"}}
         assert compact_resource(v) == "every 8h"
 
     def test_3_times_a_day(self):
+        """Test that frequency=3 over 1d period compacts to '3x/day'."""
         v = {"repeat": {"frequency": 3, "period": 1, "periodUnit": "d"}}
         assert compact_resource(v) == "3x/day"
 
     def test_3_to_4_times_a_day(self):
+        """Test that frequencyMax produces a range like '3-4x/day'."""
         v = {
             "repeat": {
                 "frequency": 3,
@@ -671,10 +800,12 @@ class TestCompactTiming:
         assert compact_resource(v) == "3-4x/day"
 
     def test_every_4_to_6_hours(self):
+        """Test that periodMax produces a range like 'every 4-6h'."""
         v = {"repeat": {"frequency": 1, "period": 4, "periodUnit": "h", "periodMax": 6}}
         assert compact_resource(v) == "every 4-6h"
 
     def test_every_21_days_for_1_hour(self):
+        """Test that duration is appended after the frequency."""
         v = {
             "repeat": {
                 "frequency": 1,
@@ -687,10 +818,12 @@ class TestCompactTiming:
         assert compact_resource(v) == "every 21d for 1h"
 
     def test_when_only(self):
+        """Test that a when-only repeat expands to the meal/event label."""
         v = {"repeat": {"when": ["CM"]}}
         assert compact_resource(v) == "at breakfast"
 
     def test_duration_with_offset_and_when(self):
+        """Test that duration, offset, and when are all combined."""
         v = {
             "repeat": {
                 "duration": 5,
@@ -704,6 +837,7 @@ class TestCompactTiming:
         )  # AC → "before meal"
 
     def test_days_of_week_with_when(self):
+        """Test that dayOfWeek and when are appended to the frequency."""
         v = {
             "repeat": {
                 "frequency": 1,
@@ -716,6 +850,7 @@ class TestCompactTiming:
         assert compact_resource(v) == "1x/day on Mon/Wed/Fri morning"
 
     def test_time_of_day(self):
+        """Test that timeOfDay is appended to the frequency."""
         v = {
             "repeat": {
                 "frequency": 1,
@@ -727,10 +862,12 @@ class TestCompactTiming:
         assert compact_resource(v) == "1x/day at 10:00"
 
     def test_count_only(self):
+        """Test that a count-only repeat compacts to 'N time(s)'."""
         v = {"repeat": {"count": 1}}
         assert compact_resource(v) == "1 time"
 
     def test_event_list(self):
+        """Test that explicit event datetimes are joined with commas."""
         v = {"event": ["2012-01-07T09:00:00+10:00", "2012-01-14T09:00:00+10:00"]}
         assert (
             compact_resource(v)
@@ -738,10 +875,12 @@ class TestCompactTiming:
         )
 
     def test_multiple_times_per_multi_day_period(self):
+        """Test that frequency over a multi-day period compacts to 'Nx/Nd'."""
         v = {"repeat": {"frequency": 2, "period": 3, "periodUnit": "d"}}
         assert compact_resource(v) == "2x/3d"
 
     def test_bounds_duration(self):
+        """Test that boundsDuration is appended as 'for N<unit>'."""
         v = {
             "repeat": {
                 "frequency": 2,
@@ -758,6 +897,7 @@ class TestCompactTiming:
         assert compact_resource(v) == "2x/day for 10d"
 
     def test_bounds_range(self):
+        """Test that boundsRange is appended as 'for low – high'."""
         v = {
             "repeat": {
                 "frequency": 3,
@@ -772,6 +912,7 @@ class TestCompactTiming:
         assert compact_resource(v) == "3x/day for 3 d – 5 d"
 
     def test_bounds_period(self):
+        """Test that boundsPeriod is appended as the compacted period string."""
         v = {
             "repeat": {
                 "frequency": 2,
@@ -783,29 +924,33 @@ class TestCompactTiming:
         assert compact_resource(v) == "2x/day from 2015-07-01"
 
     def test_when_with_frequency_no_duration(self):
+        """Test that when label is appended without duration prefix."""
         v = {
             "repeat": {"frequency": 1, "period": 1, "periodUnit": "d", "when": ["MORN"]}
         }
         assert compact_resource(v) == "1x/day morning"
 
     def test_duration_with_when_no_offset(self):
+        """Test that duration and when are combined without offset."""
         v = {"repeat": {"duration": 5, "durationUnit": "min", "when": ["AC"]}}
         assert compact_resource(v) == "for 5min before meal"
 
     def test_offset_with_when_no_duration(self):
+        """Test that offset and when are combined without duration."""
         v = {"repeat": {"offset": 10, "when": ["AC"]}}
         assert compact_resource(v) == "10min before meal"
 
     def test_empty_timing_returns_raw(self):
+        """Test that an empty Timing dict returns the raw dict."""
         v = {}
         assert compact_resource(v) == v
 
 
 class TestRealWorldPayloads:
-    # Test on an official example
-    def test_us_core_blood_pressure(self):
-        from fhir_mcp_server.compactor.dispatch import compact_resource
+    """Test compact_resource against official FHIR example payloads."""
 
+    def test_us_core_blood_pressure(self):
+        """Test compaction of the US Core Blood Pressure Observation example."""
         # The exact US Core Blood Pressure JSON payload
         payload = {
             "resourceType": "Observation",
